requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ML/Image processing
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
tweepy==4.14.0
wordfreq==3.1.1

# Development
pytest==7.4.3
black==23.11.0
ruff==0.1.6 